adds additional performance tuning found by the clr heap analyzer extension

+++ libraries/Pliant/Forest/TokenForestNode.cs
@@ -44,7 +44,7 @@ namespace Pliant.Forest
         private int ComputeHashCode()
         {
             return HashCode.Compute(
-                NodeType.GetHashCode(),
+                ((int)NodeType).GetHashCode(),
                 Location.GetHashCode(),
                 Origin.GetHashCode(),
                 Token.GetHashCode());
